- Fix class name conflicts between dll and player (e.g. class Mutex) - Fix new / delete mismatching(effectdata is correct in some case) - Fix compiler warning about numeric casting - Support PlatformMutex

+++ NativeCode/Plugin_ImpactGenerator.cpp
@@ -97,7 +97,7 @@ namespace ImpactGenerator
         {
             ImpactInstance* instance = GetImpactInstance((int)data->p[P_INSTANCE]);
             if (instance != NULL)
-                instance->maximpacts = value;
+                instance->maximpacts = (int)value;
         }
         return UNITY_AUDIODSP_OK;
     }
